pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "paperclaw"
version = "0.1.0"
description = "Turn a folder of PDFs into an organized, agent-searchable library."
requires-python = ">=3.12"
dependencies = [
    "pdfplumber>=0.11",
    "anthropic>=0.40",
    "pydantic>=2.7",
    "typer>=0.12",
    "python-dotenv>=1.0",
]

[project.scripts]
paperclaw = "paperclaw.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/paperclaw"]

[tool.ruff]
src = ["src"]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "ANN"]
ignore = ["ANN401"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN", "B017", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["paperclaw"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=paperclaw --cov-report=term-missing -q"
markers = [
    "integration: marks integration tests (deselect with '-m not integration')",
]
